Add run command
This command echoes any command passed to it and only executes it if
the DOTFILES_MODE is 'normal'. This is to facilitate development of
new commands in a safe way - if they include calls to dangerous or
long-running commands it is possible to verify that they call the expected commands in the expected order without actually running those
commands.