Add help command
This uses the specs to generate usage information for commands. This
should encourage me to write specs as a method of documenting this
repo.

This makes use of the environment variable DOTFILES_HOME which should
identify the location of the dotfiles directory on a system using it.
Temporarily, this is being initialised in the include file, but should
be done externally.