Track & install dependencies
All dependencies, no matter what installs them, are recorded in a
single file along with the command to install them. This makes it the
script to install them very easy.

Also adds a script to install a git repo based on its user & repo pair
(slash-separated as in the git url), assuming it has a 'make' task.